Revert "update the type"
This reverts commit 5cc2b0d6f5457713e418e3e0612ac01dd7836aa5.

+++ include/aws/s3/private/s3_auto_ranged_get.h
@@ -56,7 +56,7 @@ AWS_EXTERN_C_BEGIN
 AWS_S3_API struct aws_s3_meta_request *aws_s3_meta_request_auto_ranged_get_new(
     struct aws_allocator *allocator,
     struct aws_s3_client *client,
-    uint64_t part_size,
+    size_t part_size,
     const struct aws_s3_meta_request_options *options);
 
 AWS_EXTERN_C_END
