+++ source/s3_auto_ranged_put.c
@@ -227,7 +227,7 @@ static struct aws_s3_meta_request_vtable s_s3_auto_ranged_put_vtable = {
 struct aws_s3_meta_request *aws_s3_meta_request_auto_ranged_put_new(
     struct aws_allocator *allocator,
     struct aws_s3_client *client,
-    uint64_t part_size,
+    size_t part_size,
     uint64_t content_length,
     uint32_t num_parts,
     const struct aws_s3_meta_request_options *options) {
